requirements.txt
pdfplumber
langchain==0.1.20
langchain-community==0.0.38
langchain-core==0.1.52
langchain-groq==0.1.5
langchain-huggingface==0.0.3
langchain-text-splitters==0.0.2
faiss-cpu
sentence-transformers
streamlit
python-dotenv